longで受け付けるようにする

+++ AtCoderLib/整数/約数.cs
@@ -4,20 +4,15 @@ using System.Collections.Generic;
 
 class 約数
 {
-    /**
-     * <summary>
-     * 約数
-     * </summary>
-     * <param name="num">約数を求める数</param>
-     * <returns>約数の一覧</returns>
-     */
-    IEnumerable<int> GetYakusu(int n)
+
+
+    IEnumerable<long> GetYakusu(long n)
     {
-        var list = new List<int>();
-        var sqrt = (int)Math.Sqrt(n);
+        var list = new List<long>();
+        var sqrt = (long)Math.Sqrt(n);
         for (var i = 1; i <= sqrt; i++)
         {
-            var d = Math.DivRem(n, i, out int amari);
+            var d = Math.DivRem(n, i, out long amari);
             if (amari == 0)
             {
                 yield return i;
@@ -28,4 +23,6 @@ class 約数
         for (var i = list.Count - 1; i >= 0; i--)
             yield return list[i];
     }
+
+
 }
